feat(motion): update animation styles and add new transition stories
