+++ motion/_animation.scss
@@ -419,7 +419,7 @@
   }
 
   100% {
-    height: 0;
+    height: 0%;
   }
 }
