+++ packages/core/.storybook/preview.js
@@ -147,10 +147,10 @@ export const parameters = {
           'Announcements',
           ['Tegel', 'Prefix change'],
         ],
+        'Motions',
         'Foundations',
         'Utilities',
         'Patterns',
-        'Motions',
         'Components',
       ],
     },
